requirements.txt
flask
waitress
werkzeug
pdf2image
pillow
pytesseract
transformers
torch
matplotlib 
opencv-python
torchvision